+++ StripePaymentSheet/StripePaymentSheet/Source/PaymentSheet/ViewControllers/PaymentMethodFormViewController.swift
@@ -464,11 +464,15 @@ extension PaymentMethodFormViewController {
     }
 
     func handleCollectBankAccount(from viewController: UIViewController) {
-        guard
-            let usBankAccountFormElement,
-            let name = usBankAccountFormElement.name,
-            let email = usBankAccountFormElement.email
-        else {
+        guard let usBankAccountFormElement else {
+            let errorAnalytic = ErrorAnalytic(event: .unexpectedPaymentSheetError,
+                                              error: Error.usBankAccountParamsMissing)
+            STPAnalyticsClient.sharedClient.log(analytic: errorAnalytic)
+            stpAssertionFailure()
+            return
+        }
+        let billingDetails = usBankAccountFormElement.billingDetails
+        guard let name = billingDetails.name, let email = billingDetails.email else {
             let errorAnalytic = ErrorAnalytic(event: .unexpectedPaymentSheetError,
                                               error: Error.usBankAccountParamsMissing)
             STPAnalyticsClient.sharedClient.log(analytic: errorAnalytic)
@@ -478,7 +482,9 @@ extension PaymentMethodFormViewController {
 
         let params = STPCollectBankAccountParams.collectUSBankAccountParams(
             with: name,
-            email: email
+            email: email,
+            address: billingDetails.address,
+            phone: billingDetails.phone
         )
         let client = STPBankAccountCollector(style: bankAccountCollectorStyle)
         let genericError = PaymentSheetError.accountLinkFailure